C3 config

+++ app/models/tweet.rb
@@ -3,7 +3,7 @@ class Tweet
   include Mongoid::Timestamps::Created::Short #c_at
   field :text, type: String
   field :tweet_id, type: Integer
-  field :sentiment, type: Integer
+  field :sentiment, type: Float
   field :tweeted_at, type: DateTime
   field :company, type: String
 end
